#43: Prevent players from swimming up waterfalls
[Problem]

Players can swim up waterfalls in areas where we want waterfalls as decoration but also need to restrict access.

[Solution]

Handle player position packets by checking if they are in a bubble column, cancel their ability to sprint (fast swim) if so.

[Testing]

This does work partially, however there must be some discrepancy with how a user's position is tracked when they are swimming already. If you load into the given test environment, and start swimming in the normal water, then move to the bubble column, it will sometimes not cancel your swimming. We could also potentially check within a radius of the 8 blocks around the player if that's the issue.

Another potential issue with this solution is players with high ping could navigate around this by not sending a position packet until they've climbed up the waterfall on client side, and then each packet gets sent and processed but won't return them to their original location.