No Issue: Fix "Unnamed" to "Starlight" in line with name change
[Problem]

We had a lot of code references to unnamed where they should be starlight. This was also causing the test server to crash when mining blocks and the debug stick to not appear in player inventory.

[Solution]

Change all instances of "unnamed" to "starlight"

[Testing]

Tested and verified that debug stick shows up and can mine the test ores at spawn. No errors or warnings when starting up and joining server, and interacting with environment.